requirements.txt
asyncpraw
discord.py
python-dotenv
vaderSentiment
openai
